Split resolveType and react router hack into separate files
